Rename the VK service file to match the service name
